requirements.txt
streamlit
nltk
scikit-learn
pandas
numpy
PyPDF2
python-docx